:recycle: [#3489] Move mTLS/requests auth handling from stuf to soap package
The certificate handling (and configuration) is really defined on the
SOAPService which is used by the StUF service, so the latter should
only proxy the former.

Additionally, some left-over constants for StUF were still in the
SOAP package, so those have been moved into their correct locations.
