+++ src/stuf/stuf_zds/constants.py
@@ -1,6 +1,8 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
+STUF_ZDS_EXPIRY_MINUTES = 5
+
 
 class VertrouwelijkheidsAanduidingen(models.TextChoices):
     zeer_geheim = "ZEER GEHEIM", _("Zeer geheim")
